refactor(adapters): split agent_acp into acp package
- Extract usage, trajectory, client, executor, resolve, and types modules
- Point runtime/application/tests at bora.adapters.acp
- Update ARCHITECTURE tree and ACP scrape architecture gate (Refs: #31)
