+++ src/bora/adapters/acp/executor.py
@@ -1,8 +1,4 @@
-"""Single typed ACP executor (Spec 19).
-
-Parent/host only: spawns (or attaches to) an ACP entry process over stdio via
-``agent-client-protocol``. Vendor private stdout scrape is forbidden.
-"""
+"""AcpExecutor: parent-side ACP process/session invoke (Spec 19)."""
 
 from __future__ import annotations
 
@@ -11,529 +7,22 @@ import json
 import os
 import threading
 import time
-from collections.abc import Callable, Mapping, Sequence
+from collections.abc import Mapping, Sequence
 from contextlib import suppress as contextlib_suppress
-from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Any
 
 from bora import __version__ as BORA_VERSION
+from bora.adapters.acp.client import (
+    _BoraAcpClient,
+    _map_stop_reason,
+    _offline_result,
+)
+from bora.adapters.acp.trajectory import write_trajectory_jsonl
+from bora.adapters.acp.types import ProcessLauncher
+from bora.adapters.acp.usage import _as_plain_mapping, normalize_acp_usage
 from bora.adapters.acp_registry import AcpEntryDescriptor, get_entry, readiness_for
 from bora.adapters.agent_contract import AgentResult, parse_validated_text_structured
-
-# Optional process launcher: (command, args, env, cwd) -> provides streams.
-# Host default uses acp.spawn_agent_process. L1 injects docker-exec launcher.
-ProcessLauncher = Callable[..., Any]
-
-
-def _as_plain_mapping(obj: Any) -> dict[str, Any] | None:
-    """Best-effort dict from pydantic model / mapping (snake_case preferred)."""
-    if obj is None:
-        return None
-    if isinstance(obj, dict):
-        return dict(obj)
-    with contextlib_suppress(Exception):
-        if hasattr(obj, "model_dump"):
-            # exclude_none + no alias → snake_case field names from acp.schema
-            dumped = obj.model_dump(exclude_none=True)
-            if isinstance(dumped, dict):
-                return dumped
-    with contextlib_suppress(Exception):
-        if hasattr(obj, "model_dump"):
-            dumped = obj.model_dump(by_alias=True, exclude_none=True)
-            if isinstance(dumped, dict):
-                return dumped
-    return None
-
-
-def _pick_int(data: Mapping[str, Any], *keys: str) -> int | None:
-    for key in keys:
-        if key not in data:
-            continue
-        val = data[key]
-        if isinstance(val, bool):
-            continue
-        if isinstance(val, int):
-            return val
-        if isinstance(val, float) and val == int(val):
-            return int(val)
-    return None
-
-
-def _pick_number(data: Mapping[str, Any], *keys: str) -> float | int | None:
-    for key in keys:
-        if key not in data:
-            continue
-        val = data[key]
-        if isinstance(val, bool):
-            continue
-        if isinstance(val, (int, float)):
-            return val
-    return None
-
-
-def normalize_acp_usage(
-    prompt_usage: Any = None,
-    usage_update: Any = None,
-) -> dict[str, Any] | None:
-    """Merge ACP dual-source usage into one observational dict (issue #30).
-
-    Two protocol sources with different semantics:
-
-    * ``PromptResponse.usage`` (``Usage``): billable/session **token** counts
-      (``inputTokens`` / ``outputTokens`` / …).
-    * ``sessionUpdate: usage_update`` (``UsageUpdate``): **context occupancy**
-      (``used`` / ``size``) and optional session **cost** — ``used`` is *not*
-      token total.
-
-    Output shape (omit empty sections; no ``uncached_*`` fields)::
-
-        {
-          "input_tokens": int, "output_tokens": int, "total_tokens": int,
-          "thought_tokens": int, "cached_read_tokens": int, "cached_write_tokens": int,
-          "cost": {"amount": number, "currency": "USD"},
-          "context": {"used": int, "size": int},
-          "sources": {"prompt_usage": bool, "usage_update": bool}
-        }
-
-    Missing fields stay absent. Never invents tokens from ``context.used``.
-    Not PASS authority.
-    """
-    prompt_map = _as_plain_mapping(prompt_usage)
-    update_map = _as_plain_mapping(usage_update)
-    out: dict[str, Any] = {}
-    has_prompt = False
-    has_update = False
-
-    if prompt_map:
-        # Accept snake_case (model fields) and camelCase (wire / by_alias dump).
-        token_pairs = (
-            ("input_tokens", ("input_tokens", "inputTokens")),
-            ("output_tokens", ("output_tokens", "outputTokens")),
-            ("total_tokens", ("total_tokens", "totalTokens")),
-            ("thought_tokens", ("thought_tokens", "thoughtTokens")),
-            ("cached_read_tokens", ("cached_read_tokens", "cachedReadTokens")),
-            ("cached_write_tokens", ("cached_write_tokens", "cachedWriteTokens")),
-        )
-        for out_key, aliases in token_pairs:
-            val = _pick_int(prompt_map, *aliases)
-            if val is not None:
-                out[out_key] = val
-                has_prompt = True
-        # Derive total when both sides present and total missing.
-        if "total_tokens" not in out:
-            inp = out.get("input_tokens")
-            outp = out.get("output_tokens")
-            if isinstance(inp, int) and isinstance(outp, int):
-                out["total_tokens"] = inp + outp
-                has_prompt = True
-
-    if update_map:
-        # Context occupancy — never promote used → input_tokens.
-        used = _pick_int(update_map, "used")
-        size = _pick_int(update_map, "size")
-        if used is not None or size is not None:
-            ctx: dict[str, Any] = {}
-            if used is not None:
-                ctx["used"] = used
-            if size is not None:
-                ctx["size"] = size
-            out["context"] = ctx
-            has_update = True
-        cost_raw = update_map.get("cost")
-        cost_map = _as_plain_mapping(cost_raw) if cost_raw is not None else None
-        if cost_map:
-            amount = _pick_number(cost_map, "amount")
-            currency = cost_map.get("currency")
-            cost_out: dict[str, Any] = {}
-            if amount is not None:
-                cost_out["amount"] = amount
-            if isinstance(currency, str) and currency.strip():
-                cost_out["currency"] = currency.strip()
-            if cost_out:
-                out["cost"] = cost_out
-                has_update = True
-        elif used is not None or size is not None:
-            pass  # context already recorded
-        # Bare update with only sessionUpdate marker still counts as source if
-        # we already got context/cost; else ignore empty shell.
-        if not has_update and ("session_update" in update_map or "sessionUpdate" in update_map):
-            # Keep raw empty update out of normalized payload.
-            pass
-
-    if not has_prompt and not has_update:
-        return None
-
-    out["sources"] = {
-        "prompt_usage": has_prompt,
-        "usage_update": has_update,
-    }
-    return out
-
-
-def write_trajectory_jsonl(
-    inv_dir: Path,
-    *,
-    prompt: str,
-    events: tuple[dict[str, Any], ...] | list[dict[str, Any]],
-    final_text: str,
-    structured: dict[str, object] | None,
-    usage: dict[str, Any] | None,
-    ok: bool,
-    error: str | None,
-    metadata: dict[str, Any] | None = None,
-) -> Path:
-    """Write **turn-level** training trajectory for one BORA invoke.
-
-    One ``session/prompt`` (one invoke) → one training turn unit:
-
-    1. ``user`` — full prompt
-    2. ``assistant`` (optional ``part=thought``) — merged thought text
-    3. ``assistant`` — merged message text (prefer ``final_text``)
-    4. ``terminal`` — ok / error / structured / usage / stop metadata
-
-    Stream chunks are merged; raw token-level ACP updates are not written here.
-    Optional debug stream stays out of this file (see ``events.jsonl`` if needed).
-    Not PASS authority.
-    """
-    inv_dir.mkdir(parents=True, exist_ok=True)
-    path = inv_dir / "trajectory.jsonl"
-
-    thought_parts: list[str] = []
-    assistant_parts: list[str] = []
-    permission_events: list[dict[str, Any]] = []
-    acp_session_id: str | None = None
-
-    for ev in events:
-        if not isinstance(ev, dict):
-            continue
-        sid = ev.get("session_id")
-        if isinstance(sid, str) and sid:
-            acp_session_id = sid
-        if ev.get("type") == "permission_decision":
-            permission_events.append(
-                {
-                    "type": "permission_decision",
-                    "outcome": ev.get("outcome"),
-                    "option_id": ev.get("option_id"),
-                    "policy": ev.get("policy"),
-                    "source": "acp",
-                }
-            )
-            continue
-        channel = ev.get("channel")
-        text = ev.get("text")
-        if not isinstance(text, str):
-            continue
-        if channel == "thought":
-            thought_parts.append(text)
-        elif channel == "assistant":
-            assistant_parts.append(text)
-
-    merged_thought = "".join(thought_parts)
-    merged_assistant = final_text if final_text else "".join(assistant_parts)
-
-    turn_index = 1
-    if isinstance(metadata, dict):
-        # Optional: callers may pass bora turn index later.
-        raw_ti = metadata.get("turn_index")
-        if isinstance(raw_ti, int) and raw_ti > 0:
-            turn_index = raw_ti
-
-    lines: list[dict[str, Any]] = [
-        {
-            "type": "turn",
-            "role": "user",
-            "content": prompt,
-            "turn_index": turn_index,
-            "acp_session_id": acp_session_id,
-            "source": "bora",
-        }
-    ]
-    if merged_thought:
-        lines.append(
-            {
-                "type": "turn",
-                "role": "assistant",
-                "part": "thought",
-                "content": merged_thought,
-                "turn_index": turn_index,
-                "acp_session_id": acp_session_id,
-                "source": "acp",
-            }
-        )
-    lines.append(
-        {
-            "type": "turn",
-            "role": "assistant",
-            "content": merged_assistant,
-            "turn_index": turn_index,
-            "acp_session_id": acp_session_id,
-            "source": "acp",
-        }
-    )
-    for pe in permission_events:
-        pe = {**pe, "turn_index": turn_index, "acp_session_id": acp_session_id}
-        lines.append(pe)
-    lines.append(
-        {
-            "type": "terminal",
-            "ok": ok,
-            "error": error,
-            "turn_index": turn_index,
-            "acp_session_id": acp_session_id,
-            "structured": structured,
-            "usage": usage,
-            "stop_reason": (metadata or {}).get("stop_reason") if metadata else None,
-            "metadata": {
-                k: (metadata or {}).get(k)
-                for k in (
-                    "executor_kind",
-                    "acp_entry_id",
-                    "acp_version",
-                    "protocol_version",
-                    "actual_model",
-                    "locked_model",
-                )
-                if metadata and k in metadata
-            },
-            "source": "bora",
-        }
-    )
-    path.write_text(
-        "\n".join(json.dumps(x, ensure_ascii=False, sort_keys=True) for x in lines) + "\n",
-        encoding="utf-8",
-    )
-    return path
-
-
-def _offline_result(model: str) -> AgentResult:
-    return AgentResult(
-        model=model,
-        text="",
-        structured=None,
-        ok=False,
-        error="offline_forced",
-        events=(
-            {
-                "type": "lifecycle",
-                "phase": "skipped",
-                "reason": "offline_forced",
-                "source": "acp_adapter",
-            },
-        ),
-        metadata={"executor_kind": "acp"},
-    )
-
-
-def _auto_approve_permission(options: Sequence[Any]) -> Any:
-    """Batch auto-approve: select first allow-like option (Spec 19 decision 4)."""
-    import acp
-    from acp.schema import AllowedOutcome
-
-    option_id = None
-    for opt in options:
-        kind = getattr(opt, "kind", None) or (opt.get("kind") if isinstance(opt, dict) else None)
-        oid = getattr(opt, "option_id", None) or (
-            opt.get("optionId") if isinstance(opt, dict) else None
-        )
-        if kind in ("allow_once", "allow_always", "allow", "selected") or (
-            isinstance(oid, str) and "allow" in oid.lower()
-        ):
-            option_id = oid
-            break
-    if option_id is None and options:
-        first = options[0]
-        option_id = getattr(first, "option_id", None) or (
-            first.get("optionId") if isinstance(first, dict) else None
-        )
-    if not option_id:
-        option_id = "allow"
-    return acp.RequestPermissionResponse(
-        outcome=AllowedOutcome(option_id=str(option_id), outcome="selected")
-    )
-
-
-@dataclass
-class _BoraAcpClient:
-    """Minimal Client: auto-approve permission, decline elicitation, collect updates.
-
-    Usage sources are intentionally split (issue #30):
-
-    * ``latest_usage_update`` — raw ``UsageUpdate`` stream (context used/size + cost)
-    * ``prompt_usage`` — set by the executor after ``session/prompt`` returns
-      ``PromptResponse.usage`` (token counts)
-
-    Normalized merge lives in :func:`normalize_acp_usage`; do not treat
-    ``UsageUpdate.used`` as billable tokens.
-    """
-
-    events: list[dict[str, Any]] = field(default_factory=list)
-    text_chunks: list[str] = field(default_factory=list)
-    latest_usage_update: dict[str, Any] | None = None
-    prompt_usage: dict[str, Any] | None = None
-    permission_decisions: list[dict[str, Any]] = field(default_factory=list)
-    _conn: Any = None
-
-    def on_connect(self, conn: Any) -> None:
-        self._conn = conn
-
-    async def request_permission(
-        self,
-        session_id: str,
-        tool_call: Any,
-        options: list[Any],
-        **kwargs: Any,
-    ) -> Any:
-        resp = _auto_approve_permission(options)
-        decision = {
-            "type": "permission_decision",
-            "session_id": session_id,
-            "outcome": "selected",
-            "policy": "batch_auto_approve",
-            "source": "acp_client",
-        }
-        # Record option id without tool payloads that may carry paths/secrets.
-        try:
-            decision["option_id"] = resp.outcome.option_id  # type: ignore[attr-defined]
-        except Exception:  # noqa: BLE001
-            decision["option_id"] = "allow"
-        self.permission_decisions.append(decision)
-        self.events.append(decision)
-        return resp
-
-    async def session_update(self, session_id: str, update: Any, **kwargs: Any) -> None:
-        """Record full ACP stream for trajectory.jsonl (training-grade).
-
-        Intermediate AgentMessageChunk / thought text is preserved on the event
-        payload (not only text_len). Operators who run local smokes want the
-        raw stream for post-hoc training export.
-        """
-        update_type = type(update).__name__
-        event: dict[str, Any] = {
-            "type": "session_update",
-            "session_id": session_id,
-            "update_type": update_type,
-            "source": "acp",
-        }
-        # Prefer typed dump when available (full structured update).
-        with contextlib_suppress(Exception):
-            if hasattr(update, "model_dump"):
-                dumped = update.model_dump(by_alias=True, exclude_none=True)
-                if isinstance(dumped, dict):
-                    event["update"] = dumped
-
-        content = getattr(update, "content", None)
-        text = getattr(content, "text", None) if content is not None else None
-        is_agent_msg = "AgentMessage" in update_type or update_type.endswith("AgentMessageChunk")
-        is_thought = "Thought" in update_type or update_type.endswith("AgentThoughtChunk")
-        if isinstance(text, str):
-            event["text"] = text
-            event["text_len"] = len(text)
-            if is_agent_msg:
-                self.text_chunks.append(text)
-            if is_thought:
-                event["channel"] = "thought"
-            elif is_agent_msg:
-                event["channel"] = "assistant"
-        # Tool call updates: keep id/title/kind when present.
-        for attr in ("tool_call_id", "toolCallId", "title", "kind", "status"):
-            val = getattr(update, attr, None)
-            if val is not None and attr not in event:
-                # normalize camelCase keys already in model_dump; keep flat helpers
-                key = "tool_call_id" if attr in {"tool_call_id", "toolCallId"} else attr
-                event[key] = val if not hasattr(val, "value") else str(val)
-        # UsageUpdate (context occupancy + cost) — not PromptResponse token Usage.
-        dumped_raw = event.get("update")
-        dumped_update: dict[str, Any] = dumped_raw if isinstance(dumped_raw, dict) else {}
-        session_upd = (
-            getattr(update, "session_update", None)
-            or dumped_update.get("sessionUpdate")
-            or dumped_update.get("session_update")
-        )
-        is_usage_update = (
-            update_type in {"UsageUpdate", "_UsageUpdate"}
-            or "UsageUpdate" in update_type
-            or (isinstance(session_upd, str) and session_upd.lower() == "usage_update")
-        )
-        if is_usage_update:
-            raw_update = _as_plain_mapping(update)
-            if raw_update is None and isinstance(event.get("update"), dict):
-                raw_update = dict(event["update"])
-            if raw_update is not None:
-                self.latest_usage_update = raw_update
-                event["usage_update"] = raw_update
-                # Keep legacy key for older event consumers (raw UsageUpdate shape).
-                event["usage"] = raw_update
-            event["has_usage_update"] = True
-        self.events.append(event)
-
-    async def write_text_file(
-        self, session_id: str, path: str, content: str, **kwargs: Any
-    ) -> None:
-        # Not advertised; if called, refuse — tools run in entry process.
-        raise RuntimeError("acp_client_fs_proxy_disabled")
-
-    async def read_text_file(
-        self,
-        session_id: str,
-        path: str,
-        line: int | None = None,
-        limit: int | None = None,
-        **kwargs: Any,
-    ) -> Any:
-        raise RuntimeError("acp_client_fs_proxy_disabled")
-
-    async def create_terminal(self, *args: Any, **kwargs: Any) -> Any:
-        raise RuntimeError("acp_client_terminal_proxy_disabled")
-
-    async def terminal_output(self, *args: Any, **kwargs: Any) -> Any:
-        raise RuntimeError("acp_client_terminal_proxy_disabled")
-
-    async def release_terminal(self, *args: Any, **kwargs: Any) -> Any:
-        return None
-
-    async def wait_for_terminal_exit(self, *args: Any, **kwargs: Any) -> Any:
-        raise RuntimeError("acp_client_terminal_proxy_disabled")
-
-    async def kill_terminal(self, *args: Any, **kwargs: Any) -> Any:
-        return None
-
-    async def create_elicitation(self, message: str, mode: Any, **kwargs: Any) -> Any:
-        import acp
-
-        self.events.append(
-            {
-                "type": "elicitation",
-                "action": "decline",
-                "policy": "batch_decline",
-                "source": "acp_client",
-            }
-        )
-        return acp.DeclineElicitationResponse(action="decline")
-
-    async def complete_elicitation(self, elicitation_id: str, **kwargs: Any) -> None:
-        return None
-
-    async def ext_method(self, method: str, params: dict[str, Any]) -> dict[str, Any]:
-        return {}
-
-    async def ext_notification(self, method: str, params: dict[str, Any]) -> None:
-        return None
-
-
-def _map_stop_reason(stop: str | None) -> tuple[bool, str | None]:
-    if stop is None or stop in ("end_turn", "end-turn", "stop"):
-        return True, None
-    mapping = {
-        "max_tokens": "acp_stop_max_tokens",
-        "max_turn_requests": "acp_stop_max_turn_requests",
-        "refusal": "acp_stop_refusal",
-        "cancelled": "acp_cancelled",
-        "canceled": "acp_cancelled",
-    }
-    err = mapping.get(str(stop), f"acp_stop_{stop}")
-    return False, err
 
 
 class AcpExecutor:
@@ -1006,19 +495,3 @@ class AcpExecutor:
             self._conn = None
             self._process = None
             self._acp_session_id = None
-
-
-def resolve_acp_executor(
-    *,
-    entry_id: str,
-    model: str,
-    base_url: str | None = None,
-    api_key: str | None = None,
-    **_kw: Any,
-) -> AcpExecutor:
-    return AcpExecutor(
-        entry_id=entry_id,
-        model=model,
-        base_url=base_url,
-        api_key_env=api_key,
-    )
