+++ tests/acceptance/test_plugin_registry.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import os
 
-from bora.adapters.agent_acp import AcpExecutor
+from bora.adapters.acp import AcpExecutor
 from bora.adapters.agent_openai_http import OpenAIHTTPExecutor
 from bora.adapters.agent_registry import discover_executor_kinds, resolve_executor
