+++ tests/executors/test_builtin_feasibility.py
@@ -8,7 +8,7 @@ import shutil
 import pytest
 
 from bora.adapters.acp_registry import get_entry, list_entry_ids, readiness_for
-from bora.adapters.agent_acp import AcpExecutor
+from bora.adapters.acp import AcpExecutor
 from bora.adapters.executor_capabilities import BUILTIN_CAPABILITIES
